Add support for linking compartments
This adds a --compartment flag to ld.lld, which should be used instead
of -r when linking a compartment.  This  has similar behaviour to -r,
but will merge COMDATs and localise all symbols that are not exported.

This also moves $cgp address field to the middle of $cgp, doubling the
max size of $cgp. All relocations are now properly biased against it,
avoiding any temporary-out-of-bounds pointers.

No longer need to know the input section for CGP bias