Fix CI.
Cirrus CI doesn't run until there's a config file in the main branch, so
I had to commit one that doesn't work.  This PR attempts to fix it.

This runs the LLVM and clang tests.  This is very slow in CI.  These
tests take about 5 minutes on my laptop and about an hour in an 8-core
GCE container.  No idea why it's so much slower.

This includes two jobs, one that runs on PRs and runs the test suite,
the other runs on commits to the cheriot branch and uploads the built
compiler as an artefact.