[CHERIoT] Provide correct error messages for AUI{CGP,PCC}
If these are out of range, we should report an error (not crash, as we
did with AUICGP).  The AUIPCC errors were reporting the valid
relocations for AUIPC, so fix those to report the correct ones for
CHERIoT.

This is slightly wrong because we currently use CHERIoT as a CPU name
not an extension name.  We should fix that soon.

Fixes CHERIoT-Platform/llvm-project#8