[CHERIoT] Add support for large globals in some cases.
We don't have relocations for these that can support the large types,
but we do often have the size information available at compile time.
For these cases, we can emit the (possibly complex) sequence of
instructions to materialise the size and insert the CSetBounds
instruction at compile time, without involving the linker in the bounds
(the linker is still responsible for fixups in finding the base).