[cheriot] Add pre-shared objects to the linker reports.
This adds a new section that lists all of the pre-shared objects.  It
also cross-references all MMIO imports and rewrites those that point
into pre-shared objects to be SharedObject references.  These are
similar to MMIO but also contain the name of the shared object.

This cross referencing is currently quadratic and could be made faster,
but we mostly expect N to be <10, so it's not worth it yet.